test(runtime): isolate YIELD event regression
